test: Cython peephole optimizer fix (#226)
Why?
"continue" statement was being skipped during coverage.

What?
Refactor "continue" out of upload_log logic.

+++ src/djtools/sync/helpers.py
@@ -187,9 +187,11 @@ def upload_log(config: BaseConfig, log_file: Path):
     now = datetime.now()
     one_day = timedelta(days=1)
     for _file in log_file.parent.rglob("*"):
-        if _file.name == "__init__.py" or not _file.is_file():
-            continue
-        if _file.lstat().st_mtime < (now - one_day).timestamp():
+        if (
+            _file.name != "__init__.py"
+            and _file.is_file()
+            and _file.lstat().st_mtime < (now - one_day).timestamp()
+        ):
             _file.unlink()
